Fix coverage measurement for Sonarcloud.
